removed version constants from the Dict class (they're in the Versions class now) and changed version fields to int
git-svn-id: https://jtidy.svn.sourceforge.net/svnroot/jtidy/branches/CodeUpdateAndJava5@829 58265ab0-d70d-0410-8993-b38ad8f1ef24

+++ src/main/java/org/w3c/tidy/Attribute.java
@@ -78,7 +78,7 @@ public class Attribute
     /**
      * html versions for this attribute.
      */
-    private short versions;
+    private int versions;
 
     /**
      * checker for the attribute.
@@ -91,7 +91,7 @@ public class Attribute
      * @param htmlVersions versions in which this attribute is supported
      * @param check AttrCheck instance
      */
-    public Attribute(String attributeName, short htmlVersions, AttrCheck check)
+    public Attribute(String attributeName, int htmlVersions, AttrCheck check)
     {
         this.name = attributeName;
         this.versions = htmlVersions;
@@ -139,7 +139,7 @@ public class Attribute
      * @return html versions for this attribute.
      * @see Dict
      */
-    public short getVersions()
+    public int getVersions()
     {
         return this.versions;
     }
